Wait to add CLAIM command
As a transfer token will be sent in addition, the tokens will need to be persistent on restart of the shard server, and the TTL of the OFFER and token will need to be syncronized. As the bottleneck has been around receiving OFFERs and not during consignment, optimizing this to all be done in one step, can be handled at a later step. Making sure that OFFER works reliabily is the top priority.