Fix runSpaceCheck
There was an issue with this method consuming lots of memory and CPU when it was run
over large data sets. This new method uses operating system features to determine if there
is space available on the disk. This is a quick fix to address major concerns until
we have better accounting of storage and bandwidth usage.